image upload and pdf creation for invoice

+++ controllers/adminController.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 //VALIDATION
 const {validationResult} = require('express-validator/check')
@@ -38,13 +39,30 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const image = req.body.image;
+  const image = req.file;
   const description = req.body.description;
   const price = req.body.price;
   const userId = req.user;//mongoose will look for id only
   const errors = validationResult(req);
 
   if(!errors.isEmpty()){
+    return res
+    .status(422)
+    .render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing:false,
+      hasError:true,
+      product: {
+        title,
+        price,
+        description
+      },
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array()
+    });
+  }
+    if(!image){
     return res
     .status(422)
     .render('admin/edit-product', {
@@ -58,10 +76,12 @@ exports.postAddProduct = (req, res, next) => {
         price,
         description
       },
-      errorMessage: errors.array()[0].msg,
-      validationErrors: errors.array()
+      errorMessage: "Attached file is not an image (jpeg jpg or png accepted).",
+      validationErrors: []
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product({title, price, description, imageUrl, userId});
   product
@@ -106,11 +126,10 @@ exports.postEditProduct = (req, res, next)=>{
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const updatedImage = req.file;
   const updatedDescription = req.body.description;
 
   const errors = validationResult(req);
-
   if(!errors.isEmpty()){
     return res
     .status(422)
@@ -121,7 +140,6 @@ exports.postEditProduct = (req, res, next)=>{
       hasError:true,
       product: {
         title: updatedTitle,
-        imageUrl : updatedImageUrl, 
         price: updatedPrice,
         description: updatedDescription,
         _id: prodId
@@ -131,6 +149,7 @@ exports.postEditProduct = (req, res, next)=>{
     });
   }
 
+
   //calling save on existing obj updates it
   Product.findById(prodId)
   .then(product=>{
@@ -139,7 +158,11 @@ exports.postEditProduct = (req, res, next)=>{
     }
     product.title = updatedTitle;
     product.price = updatedPrice;
-    product.imageUrl = updatedImageUrl;
+    if(image){
+      //delete old image
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = image.path;
+    }
     product.description = updatedDescription;
     return product.save()  
     .then(result=>  res.redirect('/admin/products'));
@@ -154,7 +177,15 @@ exports.postEditProduct = (req, res, next)=>{
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId;
-  Product.deleteOne({_id: productId, userId: req.user._id})
+  Product
+  .findById({_id:productId})
+  .then(product=>{
+    if(!product){
+      return next(new Error('Product not found'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({_id: productId, userId: req.user._id})
+  })
   .then(result=> res.redirect('/admin/products'))
   .catch(err=>{
     const error = new Error(err);
